Refactor Application architecture using VSA
Slices:
- [x] Application
- [ ] Pages
- [x] Features
- [x] Entities
- [x] Shared
