+++ src/Application/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import reportWebVitals from './reportWebVitals';
-import App from './app/App';
+import reportWebVitals from '../Shared/reportWebVitals';
+import App from './App';
 import {Provider} from 'react-redux';
-import {store} from './app/store';
+import {store} from '../Shared/Store/store';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
