+++ src/Entities/EditableSpan/components/EditableSpan.tsx
